Create Webroot.ForceRemove.ps1
This script brute force uninstalls Webroot from a machine. It will first try the graceful methods, but immediately after that goes to brute force. It is forcefully deleting files, registry keys, and startup items...so this is to be used as a last resort if you've been previously unable to normally uninstall Webroot.